v2: consolidate dynamic-button e2e test
Drop main's v1-style dynamic-btn-upload-list.e2e.test.tsx (uses <uc-config>
sibling + dynamic-button attribute, incompatible with v2's composable model);
promote v2's smart-btn-upload-list.e2e.test.tsx to dynamic-btn-upload-list.

The broader e2e suite remains v2-spike WIP (pre-existing ~273 tsc:e2e errors,
unrelated to this rebase) and is tracked separately.
